# openapi.yaml
openapi: "3.0.0"
info:
  version: 1.0.0
  title: geeq-api
paths:
  /auth/login:
    post:
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthLoginPostRequestBody'
        required: true
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Common200Response"
          headers:
            Set-Cookie:
              schema:
                type: string
                example: geeq-session-id=abcde12345; Path=/; HttpOnly
        '401':
          description: ''
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Common401Response"
        '500':
          description: ''
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Common500Response"

  /auth/logout:
    post:
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Common200Response"
        '401':
          description: ''
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Common401Response"
        '500':
          description: ''
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Common500Response"

  /auth/me:
    get:
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthMeGet200Response"
        '401':
          description: ''
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Common401Response"
        '500':
          description: ''
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Common500Response"

components:
  schemas:
    AuthLoginPostRequestBody:
      properties:
        code:
          type: string
      required:
        - code
    AuthMeGet200Response:
      properties:
        userId:
          type: string
        avatarUrl:
          type: string
      required:
        - userId
        - avatarUrl
    Common200Response:
      properties:
        message:
          type: string
      required:
        - message
    Common401Response:
      properties:
        message:
          type: string
      required:
        - message
    Common403Response:
      properties:
        message:
          type: string
      required:
        - message
    Common500Response:
      properties:
        message:
          type: string
      required:
        - message

  # /auth/logout:
  #   post:
  # /task:
  #   get:
  # /task/{taskId}:
  #   get:
  # /task/{taskId}/submission:
  #   post:
  # /task/{taskId}/submission:
  #   get: